Synchronous callback finally done. Better UX, better UI.
Better UX, better UI: changed colors, components position, upgraded Human-Computer Interface strength and more.

One first milestone: introduced the Observer pattern in order to wait for a loading file to be ready.
Now it is possible to store the result of the class LoadMatch in a Match variable without getting a NullPointerException.
This has been achieved through the implementation of an interface and a four step Observer Pattern.

Major fixes around the code.

Added a custom font too!