Ultimated GUI
Introduction of CorruptedFileException:
now GUI handles well when there is a problem in a file user is trying to load.
Deleted old class ClickListener, because it is now simply added as actionPerformed action listener to a button.
LoadMatch fixed.

Now it is possible to load a file and play it!